+++ .idea/modules/api-server/kuddy-back.api-server.main.iml
@@ -3,4 +3,7 @@
   <component name="CheckStyle-IDEA-Module" serialisationVersion="2">
     <option name="activeLocationsIds" />
   </component>
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="bcbf9f9a-0ddc-43f7-a492-c13daffb8662" />
+  </component>
 </module>
