+++ chat-server/src/main/java/com/kuddy/chatserver/chat/controller/ChatRoomController.java
@@ -1,14 +1,12 @@
 package com.kuddy.chatserver.chat.controller;
 
-import java.io.IOException;
 import java.net.URI;
 import java.util.List;
 
 import javax.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.messaging.handler.annotation.MessageMapping;
+
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -29,15 +27,15 @@ import com.kuddy.common.security.user.AuthUser;
 import lombok.RequiredArgsConstructor;
 
 @RestController
-@RequestMapping("/chat/v1")
+@RequestMapping("/chat/v1/chatrooms")
 @RequiredArgsConstructor
-public class ChatController {
+public class ChatRoomController {
 	private final ChatService chatService;
 	private final ChatRoomService chatRoomService;
 	private static final String CHAT_ROOM_DISCONNECTED= "성공적으로 접속 해제했습니다.";
 
 
-	@PostMapping("/chatrooms")
+	@PostMapping
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> createChatRoom(@RequestBody @Valid final ChatReqDto requestDto, @AuthUser Member member) {
 
@@ -58,8 +56,20 @@ public class ChatController {
 				.build());
 	}
 
+	// 채팅방 리스트 조회
+	@GetMapping
+	@PreAuthorize("isAuthenticated()")
+	public ResponseEntity<StatusResponse> chatRoomList(@AuthUser Member member) {
+		List<ChatRoomListResDto> response = chatService.getChatList(member);
+		return ResponseEntity.ok(StatusResponse.builder()
+				.status(StatusEnum.OK.getStatusCode())
+				.message(StatusEnum.OK.getCode())
+				.data(response)
+				.build());
+	}
+
 	// 채팅내역 조회 AuthUSER
-	@GetMapping("/chatrooms/{roomId}")
+	@GetMapping("/{roomId}")
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> chattingList(@PathVariable("roomId") Long roomId, @AuthUser Member member) {
 		String email = chatService.checkRoomIdOwnerValidation(member, roomId);
@@ -72,18 +82,8 @@ public class ChatController {
 			.build());
 	}
 
-	// 채팅방 리스트 조회
-	@GetMapping("/chatrooms")
-	@PreAuthorize("isAuthenticated()")
-	public ResponseEntity<StatusResponse> chatRoomList(@AuthUser Member member) {
-		List<ChatRoomListResDto> response = chatService.getChatList(member);
-		return ResponseEntity.ok(StatusResponse.builder()
-			.status(StatusEnum.OK.getStatusCode())
-			.message(StatusEnum.OK.getCode())
-			.data(response)
-			.build());
-	}
-	@GetMapping("/chatrooms/check")
+
+	@GetMapping("/check")
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> hasChatRoom(@AuthUser Member member,@RequestParam("email") String email) {
 		Room room = chatService.findByMembers(member, email);
@@ -95,19 +95,9 @@ public class ChatController {
 				.build());
 	}
 
-	@MessageMapping("/message")
-	public void sendMessage(@Valid Message message, @Header("Authorization") final String accessToken) {
-		chatService.sendMessage(message, accessToken);
-	}
-
-	@MessageMapping("/updateMessage")
-	public void updateMessage(@Valid Message updatedMessage, @Header("Authorization") final String accessToken) throws IOException {
-		// 데이터베이스 상태 업데이트 또는 서비스 로직 실행
-		chatService.updateMessage(updatedMessage, accessToken);
-	}
 
 	// 채팅방 접속 끊기
-	@DeleteMapping("/chatrooms/{roomId}")
+	@DeleteMapping("/{roomId}")
 	@PreAuthorize("isAuthenticated()")
 	public ResponseEntity<StatusResponse> disconnectChat(@PathVariable("roomId") Long roomId,
 		@RequestParam("email") String email, @AuthUser Member member) {
